+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Setup Docker testing
   - Setup Continuous Integration testing, refs #1
   - Setup linting, refs #2
+  - Adds basic unit tests, refs #3
   - Introduce `Makefile`, refs #8
 
 ## [20190910]